Converge logger providers
