+++ src/Provider/LoggerServiceProvider.php
@@ -2,20 +2,32 @@
 
 namespace Bolt\Provider;
 
+use Bolt\Logger\Manager;
+use Bolt\Logger\Handler\SystemHandler;
 use Bolt\Logger\Handler\RecordChangeHandler;
 use Monolog\Logger;
 use Silex\Application;
 use Silex\ServiceProviderInterface;
 
 /**
- * Monolog provider for Bolt record changelog entries
+ * Monolog provider for Bolt system logging entries
  *
  * @author Gawain Lynch <gawain.lynch@gmail.com>
  */
-class LoggerRecordChangeServiceProvider implements ServiceProviderInterface
+class LoggerSystemServiceProvider implements ServiceProviderInterface
 {
     public function register(Application $app)
     {
+        // System log
+        $app['logger.system'] = $app->share(function ($app) {
+            $log = new Logger('logger.system');
+
+            $log->pushHandler(new SystemHandler($app));
+
+            return $log;
+        });
+
+        // Changelog
         $app['logger.change'] = $app->share(function ($app) {
             $log = new Logger('logger.system');
